Examine Actor: Don't Rely on 'wearing'
In `examine` ACTOR, when calculating which items are carried and which
are worn, don't check if the item is in `wearing` of the actor, just
check if it's `donned` or not, based on following assumptions:

- If it's NOT donned, it's just a carried item.
- If it's a donned clothing, it's a worn clothing item.

The above change required adding the `donned` attribute to EVERY THING,
instead of having it only on the `clothing` class.

All tests produce unchanged logs, cofirming behavior is as before.