Lib Classes: Polish Contents
The commented "Contents" section was badly indented and was causing
problems in code editors that rely on indentation for code folding.

I've polished the English and cleaned up the indentantion and line
wrapping (at col 80), and also framed the classes entries in a table-like
structure with bullet lists (this last change can easily be undone if
not liked).