ReRun Tests
Execute again all tests after the whitespace cleanup.

** NOTE **
I was hoping the trailing whitespace cleanup would solve the issue of
transcripts showing up in Git with whitespace differences at each run;
but it seems that now it's even worst (the space between two sentences
separated by a full stop is now gone). This confirms that there is a
bug in Alan relating to how it handles whitespaces after punctuation.