StdLib Man: Tuxedos
In "§6.1.1. CLOTHING" add new section title "§6.1.1.2. Clothing Tips &
Tricks" to separate contents following "§16.1.1.1. The clothing table"
from the Clothing Table and provide an independent context.

Externalize the code of the tuxedos examples to `wedding.alan`, which
now becomes an example adventure distributed to end users. Since the
full source is now available to end users, reduce the example code to
the snippets that are pertinent to the tuxedos implementation.

Fix the code in the wedding example to make it work with the latest
StdLib changes; also add a few extra tweaks to enrich the game.

Edit the contents describing the tuxedos problem and solution, and add
a transcript showing the results.

Add the `example.tuxedos` anchor ID to example code block.